pyproject.toml
[tool.poetry]
name = "trello-bot"
version = "0.1.0"
description = "Telegram bot for Trello board tracking"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.4"
colorama = "^0.4.6"
schedule = "^1.2.0"
aiohttp = "^3.9.0"
aiofiles = "^24.1.0"
aiogram = "^3.21.0"
pyyaml = "^6.0.0"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
black = "^25.1.0"
flake8 = "^7.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
